Move source/season/loss to ChargeElement model
