+++ src/lib/models/charge-element.js
@@ -1,3 +1,6 @@
+'use strict';
+const Model = require('./model');
+const { assertEnum } = require('./validators');
 
 const validSources = {
   supported: 'supported',
@@ -20,6 +23,45 @@ const validLosses = {
   nonChargeable: 'non-chargeable'
 };
 
-exports.losses = validLosses;
-exports.sources = validSources;
-exports.seasons = validSeasons;
+class ChargeElement extends Model {
+  /**
+   * Source
+   * @return {String}
+   */
+  get source () {
+    return this._source;
+  }
+
+  set source (source) {
+    assertEnum(source, Object.values(validSources));
+    this._source = source;
+  }
+
+  /**
+   * Season
+   * @return {String}
+   */
+  get season () {
+    return this._season;
+  }
+
+  set season (season) {
+    assertEnum(season, Object.values(validSeasons));
+    this._season = season;
+  }
+
+  /**
+   * Loss
+   * @return {String}
+   */
+  get loss () {
+    return this._loss;
+  }
+
+  set loss (loss) {
+    assertEnum(loss, Object.values(validLosses));
+    this._loss = loss;
+  }
+}
+
+module.exports = ChargeElement;
